move mocked_confluence_vvt into dedicated pytest plugin
